+++ CHANGES.md
@@ -4,7 +4,11 @@
 
 *  Enabled configuration of the permalink message via `config.json`, 
    with optional display of the expiration duration in days. (#500)
-
+   
+*  Simplified the AppBar by grouping features under a 
+   single button that opens a dropdown menu. The features are: 
+   Documentation, Developer Reference, Imprint, and Settings. (#540)
+   
 ### Fixes
 
 *  Applied workaround for a bug in `html-to-image` libary that causes an